requirements.txt
numpy==1.24.4
opencv-python==4.8.0.76
facenet-pytorch
mtcnn
fastapi
uvicorn
streamlit
requests
scikit-learn
torch
Pillow